% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{authors_from_bibentry}
\alias{authors_from_bibentry}
\title{Extract authors from bibentry object}
\usage{
authors_from_bibentry(be)
}
\arguments{
\item{be}{Bibentry object}
}
\value{
Character vector of author names
}
\description{
Extract authors from bibentry object
}
