% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{bioc_meta_for}
\alias{bioc_meta_for}
\title{Get Bioconductor metadata for a package without installing it}
\usage{
bioc_meta_for(pkg)
}
\arguments{
\item{pkg}{Package name}
}
\value{
List with package metadata or NULL if not found
}
\description{
Get Bioconductor metadata for a package without installing it
}
