% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{clean_author_name}
\alias{clean_author_name}
\title{Clean author names by removing bracketed content}
\usage{
clean_author_name(x)
}
\arguments{
\item{x}{Author name string}
}
\value{
Cleaned author name string
}
\description{
Clean author names by removing bracketed content
}
