% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{cran_meta_for}
\alias{cran_meta_for}
\title{Get CRAN metadata for a package without installing it}
\usage{
cran_meta_for(pkg)
}
\arguments{
\item{pkg}{Package name}
}
\value{
List with package metadata or NULL if not found
}
\description{
Get CRAN metadata for a package without installing it
}
