% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_bioc_author_from_web}
\alias{get_bioc_author_from_web}
\title{Get author information from BioC package website}
\usage{
get_bioc_author_from_web(pkg)
}
\arguments{
\item{pkg}{Package name}
}
\value{
Character string with author information or NULL if not available
}
\description{
Get author information from BioC package website
}
