% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_cran_title_from_web}
\alias{get_cran_title_from_web}
\title{Get title information from CRAN package website}
\usage{
get_cran_title_from_web(pkg)
}
\arguments{
\item{pkg}{Package name}
}
\value{
Character string with title information or NULL if not available
}
\description{
Get title information from CRAN package website
}
