% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_doi_from_github}
\alias{get_doi_from_github}
\title{Get DOI from GitHub repository}
\usage{
get_doi_from_github(repo)
}
\arguments{
\item{repo}{GitHub repository in format "owner/repo"}
}
\value{
DOI string or NULL if not found
}
\description{
Get DOI from GitHub repository
}
