% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_doi_from_package}
\alias{get_doi_from_package}
\title{Get DOI from package citation or metadata}
\usage{
get_doi_from_package(pkg)
}
\arguments{
\item{pkg}{Package name}
}
\value{
DOI string or NULL if not found
}
\description{
Get DOI from package citation or metadata
}
