% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_github_info}
\alias{get_github_info}
\title{Get GitHub repository information and DOI}
\usage{
get_github_info(pkg, url = NULL)
}
\arguments{
\item{pkg}{Package name}

\item{url}{Package URL (optional)}
}
\value{
List with GitHub info and DOI or NULL if not found
}
\description{
Get GitHub repository information and DOI
}
