% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_repository_doi}
\alias{get_repository_doi}
\title{Generate CRAN/Bioconductor DOI for a package}
\usage{
get_repository_doi(pkg, repository)
}
\arguments{
\item{pkg}{Package name}

\item{repository}{Repository name ("CRAN" or "Bioconductor")}
}
\value{
DOI string or NULL if not found
}
\description{
Generate CRAN/Bioconductor DOI for a package
}
