% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{is_bioc_package}
\alias{is_bioc_package}
\title{Check if a package is from Bioconductor}
\usage{
is_bioc_package(pkg)
}
\arguments{
\item{pkg}{Package name}
}
\value{
Logical indicating if package is from Bioconductor
}
\description{
Check if a package is from Bioconductor
}
