% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{normalize_key}
\alias{normalize_key}
\title{Normalize string for use as BibTeX key}
\usage{
normalize_key(x)
}
\arguments{
\item{x}{String to normalize}
}
\value{
Normalized string
}
\description{
Normalize string for use as BibTeX key
}
