% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{package_exists_on_repos}
\alias{package_exists_on_repos}
\title{Check if a package exists on CRAN or Bioconductor repositories}
\usage{
package_exists_on_repos(pkg)
}
\arguments{
\item{pkg}{Package name}
}
\value{
Logical indicating if package exists on any repository
}
\description{
Check if a package exists on CRAN or Bioconductor repositories
}
