% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{parse_author_text}
\alias{parse_author_text}
\title{Parse author text from DESCRIPTION file}
\usage{
parse_author_text(x)
}
\arguments{
\item{x}{Author text string}
}
\value{
Character vector of author names
}
\description{
Parse author text from DESCRIPTION file
}
