% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{person_to_string}
\alias{person_to_string}
\title{Convert person object to string format}
\usage{
person_to_string(p)
}
\arguments{
\item{p}{Person object}
}
\value{
Character string with formatted name
}
\description{
Convert person object to string format
}
