% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{ris_sanitize}
\alias{ris_sanitize}
\title{Sanitize text for RIS format (single-line)}
\usage{
ris_sanitize(x)
}
\arguments{
\item{x}{Text to sanitize}
}
\value{
Sanitized text
}
\description{
Sanitize text for RIS format (single-line)
}
