% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{safely_get}
\alias{safely_get}
\title{Safely get element from list with default value}
\usage{
safely_get(x, name, default = NULL)
}
\arguments{
\item{x}{List or object to access}

\item{name}{Name of element to get}

\item{default}{Default value if element doesn't exist}
}
\value{
Element value or default
}
\description{
Safely get element from list with default value
}
