% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{write_text}
\alias{write_text}
\title{Write text to file with overwrite control}
\usage{
write_text(path, text, overwrite = FALSE)
}
\arguments{
\item{path}{File path}

\item{text}{Text to write}

\item{overwrite}{Whether to allow overwriting existing files}
}
\description{
Write text to file with overwrite control
}
