% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{year_from_bibentry}
\alias{year_from_bibentry}
\title{Extract year from bibentry object}
\usage{
year_from_bibentry(be)
}
\arguments{
\item{be}{Bibentry object}
}
\value{
Character string with year
}
\description{
Extract year from bibentry object
}
